values for on and off
My aim for this sketch is to use within a bigger sketch to fire off the data to an automation system such as Openhab, so a value is actually required for the off state as well as the on state. Adding the publish statement to add publish the 0 value when no motion is detected, also changed the value from "Motion" to "1" when motion is detected.